Update schemas and validators
