+++ src/backend/models/plugins/wordPronunciationHook.ts
@@ -109,41 +109,3 @@ export const uploadWordPronunciation = (schema: mongoose.Schema<Interfaces.WordS
     }
   });
 };
-
-/* If the client sent over blob data for pronunciation, it will be uploaded to AWS S3 */
-export const uploadExamplePronunciation = (schema: mongoose.Schema<Interfaces.ExampleSuggestion>): void => {
-  schema.pre('save', async function (next) {
-    try {
-      if (!this.skipPronunciationHook) {
-        const id = (this._id || this.id).toString();
-
-        if (
-          // Going to mock creating and saving audio pronunciation while testing in Cypress
-          (isCypress && this.pronunciation)
-          || this.pronunciation.startsWith('data:audio/mp3')
-          || (!isCypress && !isAWSProduction && this.pronunciation)
-        ) {
-          this.pronunciation = await createAudioPronunciation(id, this.pronunciation);
-        } else if (this.pronunciation.startsWith('https://') && !this.pronunciation.includes(`${id}.`)) {
-          // If the pronunciation data for the headword is a uri, we will duplicate the uri
-          // so that the new uri will only be associated with the suggestion
-          const isMp3 = this.pronunciation.includes('mp3');
-          const oldId: string = last(compact(this.pronunciation.split(/.mp3|.webm/).join('').split('/')));
-          const newId: string = id;
-
-          /* If we are saving a new word suggestion, then we want to copy all the original audio files */
-          this.pronunciation = await (this.isNew
-            ? copyAudioPronunciation(oldId, newId, isMp3)
-            : renameAudioPronunciation(oldId, newId, isMp3));
-        }
-      }
-
-      next();
-      return this;
-    } catch (err) {
-      console.log('Error caught in pre save Example schema hook', err.message);
-      this.invalidate('pronunciation', err.message);
-      return null;
-    }
-  });
-};
